channeldb: applying payment statuses migration to current database
